feat(view): share glass capsule helper
